Improve API and docs
